+++ src/utils/mod.rs
@@ -25,5 +25,4 @@
 
 // modules
 pub mod fmt;
-pub mod hash;
 pub mod parser;